# openapi.yaml
openapi: 3.0.3
info:
  title: Hypernova Banking API
  version: 1.0.0
  license:
    name: Proprietary technical assessment implementation
    url: https://github.com/0xStyx507/hypernova-banking
  description: |
    Contrato HTTP de Hypernova Banking.

    Incluye identidad, sesiones, auditoría, cuentas USD y movimientos
    financieros respaldados por TigerBeetle. Los importes son unidades menores
    enteras serializadas como cadenas para evitar redondeos de punto flotante.
servers:
  - url: http://127.0.0.1:8080
    description: Desarrollo local
# Public health and authentication operations opt in to security below; an
# empty root requirement makes the security posture explicit for every path.
security: []
...
paths:
...
  /api/v1/auth/profile:
    put:
      operationId: updateProfile
      summary: Actualiza el nombre del usuario autenticado
      description: Requiere una sesión con MFA verificado. El correo se mantiene de solo lectura hasta implementar un flujo de verificación de cambio de email.
      tags: [Authentication]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Perfil actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Nombre inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Sesión inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: MFA no verificado en la sesión actual
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno de autenticación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Mensaje público y no sensible del error
          example: invalid credentials
        code:
          type: string
          description: Código estable para clientes
          example: insufficient_funds
...
    UpdateProfileRequest:
      type: object
      additionalProperties: false
      required: [full_name]
      properties:
        full_name:
          type: string
          minLength: 2
          maxLength: 120
          description: Nombre visible del usuario. El correo no se modifica por este endpoint.
    User:
      type: object
      required: [id, email, full_name, created_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        created_at:
          type: string
          format: date-time
...